+++ packages/spotlight/CHANGELOG.md
@@ -12,6 +12,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
+- `spotlight` to not try to focus something when the window is activated unless the window has been previously blurred
+
 ## [1.8.0] - 2017-09-07
 
 ### Fixed